requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.2
cron-descriptor==1.4.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-celery-beat==2.5.0
django-constance==3.1.0
django-extensions==3.2.3
django-picklefield==3.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.10.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.8.0
idna==3.4
inflection==0.5.1
kombu==5.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
